+++ rosbag2_snapshot/include/rosbag2_snapshot/ffmpeg_encoding/tdiff.hpp
