+++ rosbag2_snapshot/src/ffmpeg_encoding/tdiff.cpp
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "rosbag2_snapshot/tdiff.hpp"
+#include "rosbag2_snapshot/ffmpeg_encoding/tdiff.hpp"
 
 #include <iomanip>
